+++ src/ESFA.DC.KeyGenerator/KeyGenerator.cs
@@ -4,7 +4,7 @@ namespace ESFA.DC.KeyGenerator
 {
     public sealed class KeyGenerator : IKeyGenerator
     {
-        public string GenerateKey(int ukPrn, int jobId, TaskKeys task, string separator = "/")
+        public string GenerateKey(long ukPrn, long jobId, TaskKeys task, string separator = "/")
         {
             return $"{ukPrn}{separator}{jobId}{separator}{task}";
         }
